Javadoc changes
Tuning up the Blocking Observable javadocs, and settling on a consistent
set of terms throughout: Observers (not subscribers) subscribe to
Observables which invoke (not call) the Observers methods (not closures
or functions) in order to make notifications (e.g. error or complete) or
to emit (not output or send) items (not elements or objects or emissions
or values).