Deprecate manWithIndex
If we leave mapWithIndex then we should have “withIndex” on many other operators. Instead of this, an Observable should just be zipped with an infinite stream of numbers and the zip function provides the index.
This keeps the API surface area smaller and provides the desired behavior on all operators.