UnicastSubject javadoc, support for onCancelled notification.
The cancellation notification will come in handy with window (and in
groupby). It helps with the case when the main source has been cancelled
but the individual windows/groups are still being consumed. If those get
cancelled by their single Subscribers, this notification will reach the
source and once all windows/groups get cancelled, the main source can be
cancelled.

Fix to generator production accounting.