Create DESIGN.md
Initial start with terminology and types as a place to begin evolving the design of RxJava v2.

I have bootstrapped this doc with things that should not be controversial, just statements of fact of what exists. If any are controversial, I apologize, please open a PR with the recommended change.

Further proposals to this document will come via PRs for discussion and approval.