1.x: enable operator/source fusion by named operator lifter
This change factors out the body of lift() into a named class that gives
access to the operator and source parameters. By using this information,
other operators can perform what I call **operator macro-fusion**.

One such example with this PR is the repeated use of the operator
`mergeWith` which when done in the classical way creates a long
linked-list of sources merged in pairs, often leading to stack overflows
and degraded performance. However, if `mergeWith` can see that it is
applied to an existing mergeWith, the two operators can use a common
list of sources and then turn into a one-level merge() with n + 1
sources (the previous graph will then be GC'd). Don't worry, this
doesn't destroy the original assembled sequence. For example, given `c =
a.mergeWith(b); d = c.mergeWith(e);` both c and d can be freely
subscribed to and still do the same thing.

Note also that this PR conflicts with PR #3477 since the array-based
`merge(from(os))` has a different type.