1.x: split Android detection and version detection
It seems some Samsung devices running 5.x mess with field names so
reflection can't find them.

If the `PlatformDependent.resolveAndroidApiVersion` logic throws due to
the missing field (but not the class), the original code considered it
to be a non-Android platform and went ahead with `Unsafe` stuff which
causes NPEs at many places.

The change splits the class check and the version check. If the version
check fails, it returns a default version (7) value so the `Unsafe`
paths are not triggered.