Use Throwable to catch Error when get Android version
When do JUnit testing with RxRingBuffer, we may have Android platform
classes defined, but contains no Android environment.

So when we call PlatformDependent.resolveAndroidApiVersion in JUnit
testing, an UnsatisfiedLinkError will throw when calling
SystemProperties.native_get(String, String).

The UnsatisfiedLinkError is not an Exception, so the test failed because
of an uncatched exception.

Fix this by catch the Throwable instead of Exception to catch Error.