Prevent some additional memory leaks
I took a look at merging the `collect()` and `collectValuesInMemory()`
methods into a single method at the prompting of akarnokd. This made me
realize the following:
1) There are more memory leaks. I updated the setNext() method to fix some of these.
2) We can't avoid all these memory leaks without introducing schduling
because of the TimedBuffer. I don't believe engineers should have any
expectation that the TimedBuffer values will be eagerly cleaned up after
they timeout. So I left the seperate `collect()` and
`collectValuesInMemory()` methods. This allows us to test memory for
memory leaks only when we think it is appropriate.