Add createScan operator to Observable
Imagine a situation when your app has some state and wants it to be synchronized with the state of a server. The app sends its state to the server and gets back the difference between that state and the actual state of the server. Then, it applies the difference to its current state, sends the new state to the server and so on. This algorithm may be implemented using Observable.scan() operator. However, this operator seems to be redundant for the task because it acts on an upstream. In this example we will need to use a dummy Observable.range(...) stream as the upstream for Observable.scan(). The proposed operator provides a more elegant solution for the described problem and perhaps for some other problems.