Move Gradle properties into gradle.properties file
Moved version, group, and description into gradle.properties file.
Removed POM_ARTIFACT property as the 'name' property in the settings file
is used when building the POM. Removed 'release.scope' property as it
does not appear to be used. Removed VERSION_NAME property and references
to it, in favor of 'version' property.