Declare `<? super @Nullable Foo>` instead of `<@Nullable ? super Foo>`.
The latter results in runtime errors when using
`-Xnullability-annotations=@io.reactivex.rxjava3.annotations:strict`:
Kotlin interprets `? super Foo` to force the callback parameters to be
non-nullable types, apparently effectively ignoring the `@Nullable`
part?

Even if this is technically a Kotlin bug (and I don't know if it is), I
recommend `<? super @Nullable Foo>`: That is the syntax that is
preferred [by the Checker
Framework](https://checkerframework.org/manual/#annotations-on-wildcards)
and by [JSpecify](https://jspecify.dev/).

(In fact, I'd likewise recommend changing `class Foo<@NonNull T>` to
`class Foo<T extends @NonNull Object>`. In that case, Kotlin handles
both versions fine. But the latter version is better for Checker
Framework and JSpecify. I've experimented with a version of that, and
I'd be happy to send it as a pull request if you'd like to go that
direction.)